+++ src/java/Controller/LoginServlet.java
@@ -16,7 +16,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import model.Account;
-import utils.HandleMagicNumber.UserRole;
+import utils.NumberToEnum.UserRole;
 
 /**
  *
