+++ src/java/Controller/SignupServlet.java
@@ -14,7 +14,7 @@ import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import utils.HandleMagicNumber.UserRole;
+import utils.NumberToEnum.UserRole;
 import model.Account;
 
 /**
